feat(storage): implement hybrid ACK policy for S3 durability
- Add required_acks parameter to StoragePort.append() interface
- S3 adapter: acks=0 uses async buffer (existing behavior)
- S3 adapter: acks=1/-1 performs sync S3 PUT + index update before ACK
- Prevents data loss on broker crash for acks!=0 producers
- Add dedicated sync append metrics for monitoring
- Document lock ordering rules for deadlock prevention
- All 22 files updated (3 adapters + 10 mocks + 9 call sites)

Closes: hybrid-ack-policy

+++ src/infra/compression/lz4.v
@@ -207,7 +207,7 @@ fn lz4_decompress_block(compressed_data []u8) ![]u8 {
 				return error('incomplete match')
 			}
 
-			offset := int(compressed_data[ip]) | (int(u32(b) & 0xe0) << 3)
+			offset := int(u32(compressed_data[ip]) | ((u32(b) & 0xe0) << 3))
 			ip++
 
 			mut match_len := int(compressed_data[ip])
@@ -237,7 +237,7 @@ fn lz4_decompress_block(compressed_data []u8) ![]u8 {
 				return error('incomplete 3+ match')
 			}
 
-			offset := int(compressed_data[ip]) | (int(compressed_data[ip + 1]) << 8)
+			offset := int(u32(compressed_data[ip]) | (u32(compressed_data[ip + 1]) << 8))
 			ip += 2
 
 			mut match_len := int(compressed_data[ip])
